Slider in graphview working

+++ RunApproachStatistics/ViewModel/VideoViewModel.cs
@@ -468,14 +468,15 @@ namespace RunApproachStatistics.ViewModel
                             CurrentPosition = position;
                         }
                         CurrentTime = MillisecondsToTimespan(CurrentPosition);
+
                         if (videoPlaybackViewModel != null)
-            {
-                videoPlaybackViewModel.updateCurrentPosition((float)CurrentPosition / 1000);
-            }
-            else if (compareVaultsViewModel != null)
-            {
-                compareVaultsViewModel.updateCurrentPosition((float)CurrentPosition / 1000);
-            }
+                        {
+                            videoPlaybackViewModel.updateCurrentPosition(CurrentPosition / 1000);
+                        }
+                        else if (compareVaultsViewModel != null)
+                        {
+                            compareVaultsViewModel.updateCurrentPosition(CurrentPosition / 1000);
+                        }
                     }
                 }
             }
